fix: fix document preview pipeline for forked PRs (#950)
The checkout actions uses ref as `github.sha` by default. For forked
repos, this refers to the last commit on the main branch
([ref](https://docs.github.com/en/actions/reference/events-that-trigger-workflows#fork)).

This means that the docs preview pipeline when run on forked PRs does
not allow us to preview changes made in the PR.

`github.event.pull_request.head.sha` allows to to checkout the latest
commit on the forked PR.